Fix calculation of cpu usage and display processes that started after pcp-pidstat
This commit fixes the calculation of cpu usage to make it cpu=user+sys+guest. Added code to make sure processes that are started after
pcp-pidstat to be recognised. It happened because the the names from pmda needs a separate request to pmda and which has to be done explicitely.
Thus solution was to read process names as values from other metrics available and not as instance names.
As suggested by @nathans, used process instance id as key for all metrics instead of instance names and used this instance id to map the process name.