Print "?" for a calculation if a previous value for an instance is not found. Fix related tests
As a convention in PCP values that are not found/valid should be printed as "?". This commit will make the calculation method to return a None value
if any of the required previous value is not found. If a none value is returned the Reporter class would print "?" at its place.